# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/home": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "api for Admin home page",
                "operationId": "AdminHome",
                "parameters": [
                    {
                        "description": "Input Fields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginAdminData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful,welcome to home"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "api for Admin login",
                "operationId": "AdminLogin",
                "parameters": [
                    {
                        "description": "Input Fields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginAdminData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/logout": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "api for Admin logout",
                "operationId": "LogOutAdmin",
                "responses": {
                    "200": {
                        "description": "Logot successful"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "failed to send OTP"
                    }
                }
            }
        },
        "/admin/order/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin.OrderDash"
                ],
                "summary": "api for get order list",
                "operationId": "ListOrder",
                "responses": {
                    "200": {
                        "description": " get order list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AdminOrderList"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
...
        "/admin/otp-verify": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "api for Admin otp verification",
                "operationId": "OtpVerify",
                "parameters": [
                    {
                        "description": "Input Fields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OTPVerify"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "failed to send OTP"
                    }
                }
            }
        },
...
        "/admin/sales-report": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "api for admin to download sales report as csv format",
                "operationId": "SalesReport",
                "responses": {
                    "400": {
                        "description": "Missing or Invalid inputs"
                    },
                    "500": {
                        "description": "Something went wrong! failed to generate sales report"
                    }
                }
            }
        },
        "/admin/users/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.UserDash"
                ],
                "summary": "api for getuserlist",
                "operationId": "ListUsers",
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/domain.Users"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
        "/admin/users/blockuser": {
            "patch": {
                "tags": [
                    "Admin.UserDash"
                ],
                "summary": "api for admin to block or unblock user",
                "operationId": "BlockUser",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BlockUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully changed user block_status",
                        "schema": {
                            "$ref": "#/definitions/request.BlockUser"
                        }
                    },
                    "400": {
                        "description": "invalid input"
                    }
                }
            }
        },
        "/admin/users/userdetails": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.UserDash"
                ],
                "summary": "api for get user details for admin",
                "parameters": [
                    {
                        "description": "Input Fields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get user details",
                        "schema": {
                            "$ref": "#/definitions/response.UserDetails"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    }
                }
            }
        },
...
        "/user/order/cancelorder": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin.OrderDash",
                    "User.Order"
                ],
                "summary": "API for canceling an order",
                "operationId": "CancelOrder",
                "parameters": [
                    {
                        "description": "inputs",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CancelOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cancel order successful"
                    },
                    "400": {
                        "description": "Missing or invalid entry"
                    },
                    "500": {
                        "description": "Something went wrong !"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Users": {
            "type": "object",
            "required": [
                "age",
                "email",
                "first_name",
                "last_name",
                "password",
                "phone",
                "user_name"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "block_status": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 2
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 1
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 10
                },
                "updated_at": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "request.BlockUser": {
            "type": "object",
            "required": [
                "user_id",
                "user_name"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
        "request.CancelOrder": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.LoginAdminData": {
            "type": "object",
            "required": [
                "email",
                "password",
                "user_name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
...
        "request.OTPVerify": {
            "type": "object",
            "required": [
                "otp",
                "user_id"
            ],
            "properties": {
                "otp": {
                    "type": "string",
                    "maxLength": 8,
                    "minLength": 4
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.UserDetails": {
            "type": "object",
            "required": [
                "user_id",
                "user_name"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                }
            }
        },
        "response.AdminOrderList": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number"
                },
                "disc_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "net_amount": {
                    "type": "number"
                },
                "order_date": {
                    "type": "string"
                },
                "order_status_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
        "response.Cart": {
            "type": "object",
            "properties": {
                "Quantity": {
                    "type": "integer"
                },
                "Variation_name": {
                    "type": "string"
                },
                "actual_price": {
                    "type": "number"
                },
                "discount_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "qty_in_stock": {
                    "type": "integer"
                }
            }
        },
...
        "response.Orders1": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number"
                },
                "coupon_code": {
                    "type": "string"
                },
                "disc_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "description": "Use jsonb type for JSON data",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Cart"
                    }
                },
                "net_amount": {
                    "type": "number"
                },
                "order_date": {
                    "type": "string"
                },
                "order_status_name": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                }
            }
        },
...
        "response.UserDetails": {
            "type": "object",
            "properties": {
                "Orders_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Orders1"
                    }
                },
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "block_status": {
                    "type": "boolean"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "muncipality": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
    }
}